[PAGOPA-2696] chore: refactoring utility and exceptions packages
