+++ src/main/java/it/gov/pagopa/fdrtechsupport/controller/middleware/exceptionhandler/ExceptionMappers.java
@@ -5,9 +5,9 @@ import com.fasterxml.jackson.databind.JsonMappingException;
 import com.fasterxml.jackson.databind.JsonMappingException.Reference;
 import com.fasterxml.jackson.databind.exc.InvalidFormatException;
 import com.fasterxml.jackson.databind.exc.MismatchedInputException;
-import it.gov.pagopa.fdrtechsupport.exceptions.AppErrorCodeMessageEnum;
+import it.gov.pagopa.fdrtechsupport.util.error.enums.AppErrorCodeMessageEnum;
 import it.gov.pagopa.fdrtechsupport.exceptions.AppErrorCodeMessageInterface;
-import it.gov.pagopa.fdrtechsupport.exceptions.AppException;
+import it.gov.pagopa.fdrtechsupport.util.error.exception.AppException;
 import it.gov.pagopa.fdrtechsupport.models.ProblemJson;
 import jakarta.inject.Inject;
 import jakarta.validation.ConstraintViolationException;
