+++ src/main/java/it/gov/pagopa/fdrtechsupport/service/HealthCheckService.java
@@ -2,7 +2,7 @@ package it.gov.pagopa.fdrtechsupport.service;
 
 
 import it.gov.pagopa.fdrtechsupport.controller.model.response.InfoResponse;
-import it.gov.pagopa.fdrtechsupport.util.AppMessageUtil;
+import it.gov.pagopa.fdrtechsupport.util.logging.AppMessageUtil;
 import jakarta.enterprise.context.ApplicationScoped;
 import org.eclipse.microprofile.config.inject.ConfigProperty;
